+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bump Go version to 1.20
 - Plainly use the /usr/bin/time command instead of a generic time
+- Parse the `time` output for measurements represented as float64 seconds
 
 ### Fixed